+++ CHANGES.rst
@@ -4,7 +4,9 @@ Changelog
 16.4 (unreleased)
 -----------------
 
-- Nothing changed yet.
+- Rerun tests with failed subtests. This feature is only available on pytest 9.0
+  and later. The pytest-subtests plugin is *not* supported.
+  Fixes `#315 <https://github.com/pytest-dev/pytest-rerunfailures/issues/315>`_.
 
 
 16.3 (2026-05-22)